feat(be): 배달 내역 정산 관련 요약 카드 API & 정산 관리 페이지 요약 카드 API 추가 (#244)
* refactor(be): 집계 쿼리시 발생할 반환형에 알맞도록 변환

* feat(be): 상점의 일/주/월별 전체 정산 조회 및 특정 기간 조회 API 구현

* chore(be): 메서드명, 엔드포인트 애매한 이름 개선

* feat(be): 배달원의 일/주/월별 전체 정산 조회 및 특정 기간 조회 API 구현

* feat(be): 정산된 내역에서 요약 카드에 필요한 정보 조회 로직 추가

* feat(be): 정산 요약 카드 API 구현

+++ backend/src/main/java/com/deliveranything/domain/settlement/dto/UnsettledResponse.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public record UnsettledResponse(
     Long scheduledSettleAmount,
-    Integer scheduledTransactionCount
+    Long scheduledTransactionCount
 ) {
 
   public static UnsettledResponse from(List<SettlementDetail> settlementDetails) {
@@ -15,7 +15,7 @@ public record UnsettledResponse(
             .map(SettlementDetail::getTargetAmount)
             .reduce(BigDecimal.ZERO, BigDecimal::add)
             .longValue(),
-        settlementDetails.size()
+        (long) settlementDetails.size()
     );
   }
 }
